move data related tests to test package
